+++ copy_this/shopp/wp-content/shopp-addons/paymill.php
@@ -10,6 +10,11 @@
  * @since 1.2
  * @subpackage PaymillShopp
  **/
+ 
+/*
+class PaymillShopp extends GatewayFramework implements GatewayModule {
+
+*/
 
 if(defined('PAYMILL_DIR') && file_exists(PAYMILL_DIR.'lib/integration/shopplugin.inc.php')){
 	require_once(PAYMILL_DIR.'lib/integration/shopplugin.inc.php');